requirements.txt
deep_translator==1.10.1
Flask==2.2.3
genanki==0.13.0
nltk==3.8.1
numpy==1.24.2
opencv_python_headless==4.7.0.72
pytesseract==0.3.10
python-dotenv==1.0.0
spacy==3.5.2
Werkzeug==2.2.3
